requirements.txt
langchain==0.2.0
langchain_community==0.2.0
python-dotenv==1.0.1
Requests==2.32.1
streamlit==1.34.0
faiss-cpu==1.8.0
openai==1.20.0
tiktoken==0.6.0
